Moved IBM float un/packing routines out of toolkit.py into a new module.
